Neo4j storage demo -- first draft
