+++ solar/solar/core/resource/resource.py
@@ -1,28 +1,59 @@
-# -*- coding: utf-8 -*-
+from copy import deepcopy
+from multipledispatch import dispatch
 import os
 
-from copy import deepcopy
-
-import solar
-
-from solar.core import actions
-from solar.core import observer
-from solar.core import signals
-from solar.core import validation
-
-from solar.core.connections import ResourcesConnectionGraph
 from solar.interfaces.db import get_db
+from solar import utils
+
 
 db = get_db()
+
+
+# TODO: cycle detection?
+# TODO: write this as a Cypher query? Move to DB?
+def _read_input_value(input_node):
+    rel = db.get_relations(dest=input_node,
+                           type_=db.RELATION_TYPES.resource_input)
+
+    if not rel:
+        return input_node.properties.get('value')
+
+    if input_node.properties['is_list']:
+        return [_read_input_value(r.start_node) for r in rel]
+
+    return _read_input_value(rel[0].start_node)
+
+
+def prepare_meta(meta):
+    actions_path = os.path.join(meta['base_path'], 'actions')
+    meta['actions_path'] = actions_path
+    meta['base_name'] = os.path.split(meta['base_path'])[-1]
+
+    meta['actions'] = {}
+    if os.path.exists(meta['actions_path']):
+        for f in os.listdir(meta['actions_path']):
+            meta['actions'][os.path.splitext(f)[0]] = f
+
+
+def read_meta(base_path):
+    base_meta_file = os.path.join(base_path, 'meta.yaml')
+
+    metadata = utils.yaml_load(base_meta_file)
+    metadata['version'] = '1.0.0'
+    metadata['base_path'] = os.path.abspath(base_path)
+
+    return metadata
 
 
 class Resource(object):
     _metadata = {}
 
-    def __init__(self, name, metadata, args, tags=None, virtual_resource=None):
+    # Create
+    @dispatch(str, str, dict)
+    def __init__(self, name, base_path, args, tags=None, virtual_resource=None):
         self.name = name
-        if metadata:
-            self.metadata = metadata
+        if base_path:
+            self.metadata = read_meta(base_path)
         else:
             self.metadata = deepcopy(self._metadata)
 
@@ -30,208 +61,95 @@ class Resource(object):
 
         self.tags = tags or []
         self.virtual_resource = virtual_resource
+
+        self.node = db.get_or_create(
+            name,
+            args={
+                'actions_path': self.metadata.get('actions_path', ''),
+                'base_name': self.metadata.get('base_name', ''),
+                'base_path': self.metadata.get('base_path', ''),
+                'handler': self.metadata.get('handler', ''),
+                'id': self.metadata['id'],
+                'version': self.metadata.get('version', ''),
+            },
+            collection=db.COLLECTIONS.resource
+        )
+
         self.set_args_from_dict(args)
+
+    # Load
+    @dispatch(object)
+    def __init__(self, resource_node):
+        self.node = resource_node
+        self.name = resource_node.properties['name']
+        self.metadata = read_meta(resource_node.properties['base_path'])
+        self.metadata.update(resource_node.properties)
+        self.tags = []
+        self.virtual_resource = None
 
     @property
     def actions(self):
         return self.metadata.get('actions') or []
 
-    @property
-    def args(self):
-        ret = {}
+    def set_args_from_dict(self, args):
+        self.node.pull()
 
-        args = self.args_dict()
+        for k, v in self.metadata['input'].items():
+            value = args.get(k, v.get('value'))
 
-        for arg_name, metadata_arg in self.metadata['input'].items():
-            type_ = validation.schema_input_type(metadata_arg.get('schema', 'str'))
-
-            ret[arg_name] = observer.create(
-                type_, self, arg_name, args.get(arg_name)
+            i = db.get_or_create(
+                k,
+                args={
+                    'is_list': isinstance(v.get('schema'), list),
+                    'value': value,
+                },
+                collection=db.COLLECTIONS.input
+            )
+            db.get_or_create_relation(
+                self.node,
+                i,
+                args={},
+                type_=db.RELATION_TYPES.resource_input
             )
 
-        return ret
-
-    def args_dict(self):
-        raw_resource = db.read(self.name, collection=db.COLLECTIONS.resource)
-        if raw_resource is None:
-            return {}
-
-        self.metadata = raw_resource
-
-        return Resource.get_raw_resource_args(raw_resource)
-
-    def set_args_from_dict(self, new_args):
-        args = self.args_dict()
-        args.update(new_args)
-
-        self.metadata['tags'] = self.tags
-        self.metadata['virtual_resource'] = self.virtual_resource
-        for k, v in args.items():
-            if k not in self.metadata['input']:
-                raise NotImplementedError(
-                    'Argument {} not implemented for resource {}'.format(k, self)
-                )
-
-            if isinstance(v, dict) and 'value' in v:
-                v = v['value']
-            self.metadata['input'][k]['value'] = v
-
-        db.save(self.name, self.metadata, collection=db.COLLECTIONS.resource)
-
-    def set_args(self, args):
-        self.set_args_from_dict({k: v.value for k, v in args.items()})
-
-    def __repr__(self):
-        return ("Resource(name='{id}', metadata={metadata}, args={input}, "
-                "tags={tags})").format(**self.to_dict())
-
-    def color_repr(self):
-        import click
-
-        arg_color = 'yellow'
-
-        return ("{resource_s}({name_s}='{id}', {metadata_s}={metadata}, "
-                "{args_s}={input}, {tags_s}={tags})").format(
-            resource_s=click.style('Resource', fg='white', bold=True),
-            name_s=click.style('name', fg=arg_color, bold=True),
-            metadata_s=click.style('metadata', fg=arg_color, bold=True),
-            args_s=click.style('args', fg=arg_color, bold=True),
-            tags_s=click.style('tags', fg=arg_color, bold=True),
-            **self.to_dict()
-        )
-
-    def to_dict(self):
+    @property
+    def args(self):
         return {
-            'id': self.name,
-            'metadata': self.metadata,
-            'input': self.args_show(),
-            'tags': self.tags,
+            k: n.properties['value']
+            for k, n in self.resource_inputs().items()
         }
 
-    def args_show(self):
-        def formatter(v):
-            if isinstance(v, observer.ListObserver):
-                return v.value
-            elif isinstance(v, observer.Observer):
-                return {
-                    'emitter': v.emitter.attached_to.name if v.emitter else None,
-                    'value': v.value,
-                }
-
-            return v
-
-        return {k: formatter(v) for k, v in self.args.items()}
-
-    def add_tag(self, tag):
-        if tag not in self.tags:
-            self.tags.append(tag)
-
-    def remove_tag(self, tag):
-        try:
-            self.tags.remove(tag)
-        except ValueError:
-            pass
-
-    def notify(self, emitter):
-        """Update resource's args from emitter's args.
-
-        :param emitter: Resource
-        :return:
-        """
-        r_args = self.args
-
-        for key, value in emitter.args.iteritems():
-            r_args[key].notify(value)
-
     def update(self, args):
-        """This method updates resource's args with a simple dict.
+        # TODO: disconnect input when it is updated and and end_node
+        #       for some input_to_input relation
+        resource_inputs = self.resource_inputs()
 
-        :param args:
-        :return:
-        """
-        # Update will be blocked if this resource is listening
-        # on some input that is to be updated -- we should only listen
-        # to the emitter and not be able to change the input's value
-        r_args = self.args
+        for k, v in args.items():
+            i = resource_inputs[k]
+            i.properties['value'] = v
+            i.push()
 
-        for key, value in args.iteritems():
-            r_args[key].update(value)
+    def resource_inputs(self):
+        inputs =  [
+            r.end_node for r in
+            db.get_relations(source=self.node,
+                             type_=db.RELATION_TYPES.resource_input)
+        ]
 
-        self.set_args(r_args)
-
-    def action(self, action):
-        if action in self.actions:
-            actions.resource_action(self, action)
-        else:
-            raise Exception('Uuups, action is not available')
-
-    @staticmethod
-    def get_raw_resource_args(raw_resource):
-        return {k: v.get('value') for k, v in raw_resource['input'].items()}
+        return {
+            i.properties['name']: i for i in inputs
+        }
 
 
-def wrap_resource(raw_resource):
-    name = raw_resource['id']
-    args = Resource.get_raw_resource_args(raw_resource)
-    tags = raw_resource.get('tags', [])
-    virtual_resource = raw_resource.get('virtual_resource', [])
+def load(name):
+    r = db.get(name, collection=db.COLLECTIONS.resource)
 
-    return Resource(name, raw_resource, args, tags=tags, virtual_resource=virtual_resource)
+    if not r:
+        raise Exception('Resource {} does not exist in DB'.format(name))
 
-
-def wrap_resource_no_value(raw_resource):
-    name = raw_resource['id']
-    args = {k: v for k, v in raw_resource['input'].items()}
-    tags = raw_resource.get('tags', [])
-    virtual_resource = raw_resource.get('virtual_resource', [])
-
-    return Resource(name, raw_resource, args, tags=tags, virtual_resource=virtual_resource)
+    return wrap_resource(r)
 
 
-def load(resource_name):
-    raw_resource = db.read(resource_name, collection=db.COLLECTIONS.resource)
+def wrap_resource(resource_node):
+    return Resource(resource_node)
 
-    if raw_resource is None:
-        raise KeyError(
-            'Resource {} does not exist'.format(resource_name)
-        )
-
-    return wrap_resource(raw_resource)
-
-
-def load_all():
-    ret = {}
-
-    for raw_resource in db.get_list(collection=db.COLLECTIONS.resource):
-        resource = wrap_resource(raw_resource)
-        ret[resource.name] = resource
-
-    return ret
-
-
-def assign_resources_to_nodes(resources, nodes):
-    for node in nodes:
-        for resource in resources:
-            res = deepcopy(resource)
-            res['tags'] = list(set(node.get('tags', [])) |
-                               set(resource.get('tags', [])))
-            resource_uuid = '{0}-{1}'.format(res['id'], solar.utils.generate_uuid())
-            # We should not generate here any uuid's, because
-            # a single node should be represented with a single
-            # resource
-            node_uuid = node['id']
-
-            node_resource_template = solar.utils.read_config()['node_resource_template']
-            args = {k: v['value'] for k, v in res['input'].items()}
-            created_resource = create(resource_uuid, resource['dir_path'], args, tags=res['tags'])
-            created_node = create(node_uuid, node_resource_template, node, tags=node.get('tags', []))
-
-            signals.connect(created_node, created_resource)
-
-
-def connect_resources(profile):
-    connections = profile.get('connections', [])
-    graph = ResourcesConnectionGraph(connections, load_all().values())
-
-    for connection in graph.iter_connections():
-        signals.connect(connection['from'], connection['to'], connection['mapping'])
