+++ solar/solar/core/resource_old/__init__.py
@@ -1,0 +1,22 @@
+__all__ = [
+    'Resource',
+    'assign_resources_to_nodes',
+    'connect_resources',
+    'create',
+    'load',
+    'load_all',
+    'prepare_meta',
+    'wrap_resource',
+    'validate_resources',
+]
+
+
+from solar.core.resource.resource import Resource
+from solar.core.resource.resource import assign_resources_to_nodes
+from solar.core.resource.resource import connect_resources
+from solar.core.resource.resource import load
+from solar.core.resource.resource import load_all
+from solar.core.resource.resource import wrap_resource
+from solar.core.resource.virtual_resource import create
+from solar.core.resource.virtual_resource import prepare_meta
+from solar.core.resource.virtual_resource import validate_resources
