+++ solar/solar/core/resource_old/compiler.py
