+++ solar/solar/interfaces/db/__init__.py
@@ -1,8 +1,10 @@
 
+from solar.interfaces.db.neo4j import Neo4jDB
 from solar.interfaces.db.redis_db import RedisDB
 from solar.interfaces.db.redis_db import FakeRedisDB
 
 mapping = {
+    'neo4j_db': Neo4jDB,
     'redis_db': RedisDB,
     'fakeredis_db': FakeRedisDB
 }
@@ -14,5 +16,5 @@ def get_db():
     # Should be retrieved from config
     global DB
     if DB is None:
-        DB = mapping['redis_db']()
+        DB = mapping['neo4j_db']()
     return DB
