+++ solar/solar/test/base.py
@@ -5,7 +5,6 @@ import unittest
 import yaml
 
 from solar.core.resource import virtual_resource as vr
-from solar.core import signals as xs
 from solar.interfaces.db import get_db
 
 db = get_db()
@@ -18,7 +17,6 @@ class BaseResourceTest(unittest.TestCase):
     def tearDown(self):
         shutil.rmtree(self.storage_dir)
         db.clear()
-        xs.Connections.clear()
 
     def make_resource_meta(self, meta_yaml):
         meta = yaml.load(meta_yaml)
